refactor: GameMachine 수정
